emergency update

+++ src/models/refundModel.js
@@ -3,110 +3,6 @@ const sequelize=require("../db/db")
 const Sequelize=require("sequelize")
 
 
-// const refund=sequelize.define("refund", {
-//     refundId:{
-//         type:Sequelize.UUID,
-//         defaultValue:Sequelize.UUIDV4,
-//         allowNull:false,
-//         primaryKey:true
-//     },
-//     transectionReference:{
-//         type:Sequelize.STRING,
-//         allowNull:false,
-//         // primaryKey:true
-//     },
-//     userId:{
-//         type:Sequelize.UUID,
-//         allowNull:false,
-//     },
-//     complainId:{
-//         type:Sequelize.UUID,
-//         allowNull:false,
-//     },
-//     debitAccount:{
-//         type:Sequelize.STRING,
-//         allowNull:false,
-//     },
-//     creditAccount:{
-//         type:Sequelize.STRING,
-//         allowNull:false
-//     },
-//     amount:{
-//         type:Sequelize.FLOAT,
-//         allowNull:false
-//     },
-//     remark:{
-//         type:Sequelize.TEXT,
-//         allowNull:false
-//     },
-//     issueDate:{
-//         type:Sequelize.DATEONLY,
-//         allowNull:false
-//     },
-//     transectionDate:{
-//         type: Sequelize.DATEONLY,
-//         allowNull: true,
-//         defaultValue: Sequelize.NOW, // Default value to current date
-//     }
-    
-// })
-
-// module.exports=refund
-
-// const refund = sequelize.define("refund", {
-//     refundId: {
-//         type: Sequelize.UUID,
-//         defaultValue: Sequelize.UUIDV4,
-//         allowNull: false,
-//         primaryKey: true
-//     },
-//     transectionReference: {
-//         type: Sequelize.STRING,
-//         allowNull: false,
-//     },
-//     userId: {
-//         type: Sequelize.UUID,
-//         allowNull: false,
-//     },
-//     complainId: {
-//         type: Sequelize.UUID,
-//         allowNull: false, // Foreign key
-//     },
-//     debitAccount: {
-//         type: Sequelize.STRING,
-//         allowNull: false,
-//     },
-//     creditAccount: {
-//         type: Sequelize.STRING,
-//         allowNull: false,
-//     },
-//     amount: {
-//         type: Sequelize.FLOAT,
-//         allowNull: false,
-//     },
-//     remark: {
-//         type: Sequelize.TEXT,
-//         allowNull: false,
-//     },
-//     issueDate: {
-//         type: Sequelize.DATEONLY,
-//         allowNull: false,
-//     },
-//     transectionDate: {
-//         type: Sequelize.DATEONLY,
-//         allowNull: true,
-//         defaultValue: Sequelize.NOW,
-//     }
-// });
-
-// refund.belongsTo(require("./reconciliation"), { foreignKey: "complainId", targetKey: "complainId" }); // Association
-// module.exports = refund;
-
-
-
-
-
-
 
 // const sequelize = require("../db/db");
 // const Sequelize = require("sequelize");
@@ -148,8 +44,12 @@ const refund = sequelize.define("refund", {
         type:Sequelize.STRING,
         allowNull:false
     },
+    // amount: {
+    //     type: Sequelize.FLOAT,
+    //     allowNull: false,
+    // },
     amount: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(15,2),
         allowNull: false,
     },
     remark: {
